requirements.txt
pandas>=1.5.0
numpy>=1.20.0
requests>=2.28.0
aiohttp>=3.8.0
finta>=1.3
python-dateutil>=2.8.0
python-dotenv>=0.19.0
scipy>=1.9.0
scikit-learn>=1.1.0
tenacity>=8.0.0
colorlog>=6.6.0
rich>=12.0.0
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
cryptography>=3.4.0
pytz>=2021.1